# Fixed bug in inc_rp_optimized.cl on Apple Intel with Metal