--- a/OpenCL/inc_rp_optimized.h
+++ b/OpenCL/inc_rp_optimized.h
@@ -14,7 +14,7 @@
 #define MAYBE_UNUSED
 #endif
 
-#ifdef IS_METAL
+#if defined(IS_METAL) && !defined(IS_APPLE_SILICON)
 #define HC_INLINE_RP __attribute__ ((noinline))
 #else
 #define HC_INLINE_RP