--- a/docs/changes.txt
+++ b/docs/changes.txt
@@ -91,6 +91,7 @@
 - Added verification of token buffer length when using TOKEN_ATTR_FIXED_LENGTH
 - Fixed a bug in all SCRYPT-based hash modes with Apple Metal
 - Fixed buffer overflow on module_26600.c / module_hash_encode()
+- Fixed bug in inc_rp_optimized.cl on Apple Intel with Metal
 - Fixed bug in module_constraints and kernel for hash-mode 7801
 - Fixed bug in module_constraints and kernel for hash-mode 7800
 - Fixed bug in 18400 module_hash_encode
@@ -175,9 +176,10 @@
 - Metal Backend: added workaround to set the true Processor value in Metal devices on Apple Intel
 - Metal Backend: added support to 2D/3D Compute
 - Metal Backend: allow use of devices with Metal if runtime version is >= 200
+- Metal Backend: define USE_BITSELECT and USE_SWIZZLE for Apple Intel
 - Metal Backend: disable Metal devices only if at least one OpenCL device is active
 - Metal Backend: improved compute workloads calculation
-- Metal Backend: define USE_BITSELECT and USE_SWIZZLE for Apple Intel
+- Metal Backend: parallelize pipeline state object (PSO) compilation internally
 - Modules: Check UnpackSize to raise false positive with hc_decompress_rar
 - User Options: added --metal-compiler-runtime option
 - User Options: limit --bitmap-max value to 31