@jsteube 2025-07-14
Benchmark with 50m random pws, single hash ntlm, passwordspro rule, nv 4090

optimized:
real    0m10,549s
user    0m1,614s
sys     0m0,354s

pure:
real    1m17,226s
user    0m1,831s
sys     0m0,251s

After patch:

optimized:
real    0m12,556s
user    0m1,600s
sys     0m0,360s

pure:
real    1m17,228s
user    0m2,902s
sys     0m0,251s

I ran all commands multiple times to make sure all caching is in place, and fixed -n -T and -u.

-> Speed loss around 18%

We can see no change in pure kernel, makes sense, since pure kernels have not been changed.

I tried another GPU, here a 1080Ti, now in MH/s

Original: 490 MH/s
Patched: 420 MH/s